+++ packages/backend/src/core/entities/DriveFileEntityService.ts
@@ -13,6 +13,7 @@ import { deepClone } from '@/misc/clone.js';
 import { UtilityService } from '../UtilityService.js';
 import { UserEntityService } from './UserEntityService.js';
 import { DriveFolderEntityService } from './DriveFolderEntityService.js';
+import { VideoProcessingService } from '../VideoProcessingService.js';
 
 type PackOptions = {
 	detail?: boolean,
@@ -20,6 +21,7 @@ type PackOptions = {
 	withUser?: boolean,
 };
 import { bindThis } from '@/decorators.js';
+import { isMimeImage } from '@/misc/is-mime-image.js';
 
 @Injectable()
 export class DriveFileEntityService {
@@ -42,6 +44,7 @@ export class DriveFileEntityService {
 
 		private utilityService: UtilityService,
 		private driveFolderEntityService: DriveFolderEntityService,
+		private videoProcessingService: VideoProcessingService,
 	) {
 	}
 	
@@ -71,40 +74,64 @@ export class DriveFileEntityService {
 	}
 
 	@bindThis
-	public getPublicUrl(file: DriveFile, mode? : 'static' | 'avatar'): string | null { // static = thumbnail
-		const proxiedUrl = (url: string) => appendQuery(
+	private getProxiedUrl(url: string, mode?: 'static' | 'avatar'): string | null {
+		return appendQuery(
 			`${this.config.mediaProxy}/${mode ?? 'image'}.webp`,
 			query({
 				url,
 				...(mode ? { [mode]: '1' } : {}),
 			})
-		);
+		)
+	}
 
-		// リモートかつメディアプロキシ
-		if (file.uri != null && file.userHost != null && this.config.externalMediaProxyEnabled) {
-			return proxiedUrl(file.uri);
-		}
+	@bindThis
+	public getThumbnailUrl(file: DriveFile): string | null {
+		if (file.type.startsWith('video')) {
+			if (file.thumbnailUrl) return file.thumbnailUrl;
 
-		// リモートかつ期限切れはローカルプロキシを試みる
-		if (file.uri != null && file.isLink && this.config.proxyRemoteFiles) {
-			const key = mode === 'static' ? file.thumbnailAccessKey : file.webpublicAccessKey;
-
-			if (key && !key.match('/')) {	// 古いものはここにオブジェクトストレージキーが入ってるので除外
-				const url = `${this.config.url}/files/${key}`;
-				if (mode === 'avatar') return proxiedUrl(url);
-				return url;
+			if (this.config.videoThumbnailGenerator == null) {
+				return this.videoProcessingService.getExternalVideoThumbnailUrl(file.webpublicUrl ?? file.url ?? file.uri);
 			}
+		} else if (file.uri != null && file.userHost != null && this.config.externalMediaProxyEnabled) {
+			// 動画ではなくリモートかつメディアプロキシ
+			return this.getProxiedUrl(file.uri, 'static');
 		}
 
-		const isImage = file.type && ['image/png', 'image/apng', 'image/gif', 'image/jpeg', 'image/webp', 'image/avif', 'image/svg+xml'].includes(file.type);
-
-		if (mode === 'static') {
-			return file.thumbnailUrl ?? (isImage ? (file.webpublicUrl ?? file.url) : null);
+		if (file.uri != null && file.isLink && this.config.proxyRemoteFiles) {
+			// リモートかつ期限切れはローカルプロキシを試みる
+			// 従来は/files/${thumbnailAccessKey}にアクセスしていたが、
+			// /filesはメディアプロキシにリダイレクトするようにしたため直接メディアプロキシを指定する
+			return this.getProxiedUrl(file.uri, 'static');
 		}
 
 		const url = file.webpublicUrl ?? file.url;
 
-		if (mode === 'avatar') return proxiedUrl(url);
+		return file.thumbnailUrl ?? (isMimeImage(file.type, 'sharp-convertible-image') ? this.getProxiedUrl(url, 'static') : null);
+	}
+
+	@bindThis
+	public getPublicUrl(file: DriveFile, mode?: 'avatar'): string | null { // static = thumbnail
+		// リモートかつメディアプロキシ
+		if (file.uri != null && file.userHost != null && this.config.externalMediaProxyEnabled) {
+			return this.getProxiedUrl(file.uri, mode);
+		}
+
+		// リモートかつ期限切れはローカルプロキシを試みる
+		if (file.uri != null && file.isLink && this.config.proxyRemoteFiles) {
+			const key = file.webpublicAccessKey;
+
+			if (key && !key.match('/')) {	// 古いものはここにオブジェクトストレージキーが入ってるので除外
+				const url = `${this.config.url}/files/${key}`;
+				if (mode === 'avatar') return this.getProxiedUrl(file.uri, 'avatar');
+				return url;
+			}
+		}
+
+		const url = file.webpublicUrl ?? file.url;
+
+		if (mode === 'avatar') {
+			return this.getProxiedUrl(url, 'avatar');
+		}
 		return url;
 	}
 
@@ -181,7 +208,7 @@ export class DriveFileEntityService {
 			blurhash: file.blurhash,
 			properties: opts.self ? file.properties : this.getPublicProperties(file),
 			url: opts.self ? file.url : this.getPublicUrl(file),
-			thumbnailUrl: this.getPublicUrl(file, 'static'),
+			thumbnailUrl: this.getThumbnailUrl(file),
 			comment: file.comment,
 			folderId: file.folderId,
 			folder: opts.detail && file.folderId ? this.driveFolderEntityService.pack(file.folderId, {
@@ -216,7 +243,7 @@ export class DriveFileEntityService {
 			blurhash: file.blurhash,
 			properties: opts.self ? file.properties : this.getPublicProperties(file),
 			url: opts.self ? file.url : this.getPublicUrl(file),
-			thumbnailUrl: this.getPublicUrl(file, 'static'),
+			thumbnailUrl: this.getThumbnailUrl(file),
 			comment: file.comment,
 			folderId: file.folderId,
 			folder: opts.detail && file.folderId ? this.driveFolderEntityService.pack(file.folderId, {