+++ packages/backend/src/server/FileServerService.ts
@@ -146,8 +146,16 @@ export class FileServerService {
 						const url = new URL(`${this.config.mediaProxy}/static.webp`);
 						url.searchParams.set('url', file.url);
 						url.searchParams.set('static', '1');
+
+						file.cleanup();
 						return await reply.redirect(301, url.toString());
 					} else if (file.mime.startsWith('video/')) {
+						const externalThumbnail = this.videoProcessingService.getExternalVideoThumbnailUrl(file.url);
+						if (externalThumbnail) {
+							file.cleanup();
+							return await reply.redirect(301, externalThumbnail);
+						}
+
 						image = await this.videoProcessingService.generateVideoThumbnail(file.path);
 					}
 				}
@@ -158,6 +166,8 @@ export class FileServerService {
 
 						const url = new URL(`${this.config.mediaProxy}/svg.webp`);
 						url.searchParams.set('url', file.url);
+
+						file.cleanup();
 						return await reply.redirect(301, url.toString());
 					}
 				}
@@ -251,8 +261,21 @@ export class FileServerService {
 			const isConvertibleImage = isMimeImage(file.mime, 'sharp-convertible-image');
 			const isAnimationConvertibleImage = isMimeImage(file.mime, 'sharp-animation-convertible-image');
 
+			if (
+				'emoji' in request.query ||
+				'avatar' in request.query ||
+				'static' in request.query ||
+				'preview' in request.query ||
+				'badge' in request.query
+			) {
+				if (!isConvertibleImage) {
+					// 画像でないなら404でお茶を濁す
+					throw new StatusError('Unexpected mime', 404);
+				}
+			}
+
 			let image: IImageStreamable | null = null;
-			if (('emoji' in request.query || 'avatar' in request.query) && isConvertibleImage) {
+			if ('emoji' in request.query || 'avatar' in request.query) {
 				if (!isAnimationConvertibleImage && !('static' in request.query)) {
 					image = {
 						data: fs.createReadStream(file.path),
@@ -273,16 +296,11 @@ export class FileServerService {
 						type: 'image/webp',
 					};
 				}
-			} else if ('static' in request.query && isConvertibleImage) {
+			} else if ('static' in request.query) {
 				image = this.imageProcessingService.convertToWebpStream(file.path, 498, 280);
-			} else if ('preview' in request.query && isConvertibleImage) {
+			} else if ('preview' in request.query) {
 				image = this.imageProcessingService.convertToWebpStream(file.path, 200, 200);
 			} else if ('badge' in request.query) {
-				if (!isConvertibleImage) {
-					// 画像でないなら404でお茶を濁す
-					throw new StatusError('Unexpected mime', 404);
-				}
-
 				const mask = sharp(file.path)
 					.resize(96, 96, {
 						fit: 'inside',