+++ packages/backend/test/utils.ts
@@ -11,7 +11,7 @@ import FormData from 'form-data';
 import { DataSource } from 'typeorm';
 import got, { RequestError } from 'got';
 import loadConfig from '../src/config/load.js';
-import { entities } from '../src/postgre.js';
+import { entities } from '@/postgres.js';
 import type * as misskey from 'misskey-js';
 
 const _filename = fileURLToPath(import.meta.url);